+++ wemake_python_styleguide/logic/tree/functions.py
@@ -10,20 +10,21 @@ from wemake_python_styleguide.types import (
 )
 
 
-def given_function_called(node: Call, to_check: Container[str]) -> str:
+def given_function_called(
+    node: Call,
+    to_check: Container[str],
+    *,
+    split_modules: bool = False,
+) -> str:
     """
     Returns function name if it is called and contained in the container.
 
-    >>> import ast
-    >>> module = ast.parse('print(123, 456)')
-    >>> given_function_called(module.body[0].value, ['print'])
-    'print'
-
-    >>> given_function_called(module.body[0].value, ['adjust'])
-    ''
-
+    If `split_modules`, takes the modules or objects into account. Otherwise,
+    it only cares about the function's name.
     """
     function_name = source.node_to_string(node.func)
+    if split_modules:
+        function_name = function_name.split('.')[-1]
     if function_name in to_check:
         return function_name
     return ''