+++ wemake_python_styleguide/visitors/ast/builtins.py
@@ -15,7 +15,7 @@ from typing import (
 )
 from typing.re import Pattern
 
-from typing_extensions import final
+from typing_extensions import Final, final
 
 from wemake_python_styleguide import constants
 from wemake_python_styleguide.compat.aliases import (
@@ -25,10 +25,19 @@ from wemake_python_styleguide.compat.aliases import (
 )
 from wemake_python_styleguide.logic import nodes, safe_eval, source, walk
 from wemake_python_styleguide.logic.naming.name_nodes import extract_name
-from wemake_python_styleguide.logic.tree import operators, strings
-from wemake_python_styleguide.logic.tree.functions import given_function_called
-from wemake_python_styleguide.types import AnyFor, AnyNodes, AnyText, AnyWith
-from wemake_python_styleguide.violations import best_practices, consistency
+from wemake_python_styleguide.logic.tree import attributes, operators, strings
+from wemake_python_styleguide.types import (
+    AnyChainable,
+    AnyFor,
+    AnyNodes,
+    AnyText,
+    AnyWith,
+)
+from wemake_python_styleguide.violations import (
+    best_practices,
+    complexity,
+    consistency,
+)
 from wemake_python_styleguide.visitors import base, decorators
 
 #: Items that can be inside a hash.
@@ -90,17 +99,6 @@ class WrongStringVisitor(base.BaseNodeVisitor):
         self._check_modulo_patterns(node, text_data)
         self.generic_visit(node)
 
-    def visit_JoinedStr(self, node: ast.JoinedStr) -> None:
-        """
-        Forbids to use ``f`` strings.
-
-        Raises:
-            FormattedStringViolation
-
-        """
-        self.add_violation(consistency.FormattedStringViolation(node))
-        self.generic_visit(node)
-
     def _check_is_alphatbet(
         self,
         node: AnyText,
@@ -145,6 +143,116 @@ class WrongStringVisitor(base.BaseNodeVisitor):
                 self.add_violation(
                     consistency.ModuloStringFormatViolation(node),
                 )
+
+
+@final
+class WrongFormatStringVisitor(base.BaseNodeVisitor):
+    """Restricts usage of ``f`` strings."""
+
+    _valid_format_index: ClassVar[AnyNodes] = (
+        *TextNodes,
+        ast.Num,
+        ast.Name,
+        ast.NameConstant,
+    )
+    _single_use_types: ClassVar[AnyNodes] = (
+        ast.Call,
+        ast.Subscript,
+    )
+    _chainable_types: Final = (
+        ast.Call,
+        ast.Subscript,
+        ast.Attribute,
+    )
+    _max_chained_items = 3
+
+    def visit_JoinedStr(self, node: ast.JoinedStr) -> None:
+        """
+        Forbids use of ``f`` strings and too complex ``f`` strings.
+
+        Raises:
+            FormattedStringViolation
+            TooComplexFormattedStringViolation
+
+        """
+        self._check_complex_formatted_string(node)
+
+        # We don't allow `f` strings by default:
+        self.add_violation(consistency.FormattedStringViolation(node))
+        self.generic_visit(node)
+
+    def _check_complex_formatted_string(self, node: ast.JoinedStr) -> None:
+        """
+        Whitelists all simple uses of f strings.
+
+        Checks if list, dict, function call with no parameters or variable.
+        """
+        has_formatted_components = any(
+            isinstance(comp, ast.FormattedValue)
+            for comp in node.values
+        )
+        if not has_formatted_components:
+            self.add_violation(  # If no formatted values
+                complexity.TooComplexFormattedStringViolation(node),
+            )
+            return
+
+        for string_component in node.values:
+            if isinstance(string_component, ast.FormattedValue):
+                # Test if possible chaining is invalid
+                format_value = string_component.value
+                if self._is_valid_formatted_value(format_value):
+                    continue
+                self.add_violation(  # Everything else is too complex:
+                    complexity.TooComplexFormattedStringViolation(node),
+                )
+                break
+
+    def _is_valid_formatted_value(self, format_value: ast.AST) -> bool:
+        if isinstance(format_value, self._chainable_types):
+            if not self._is_valid_chaining(format_value):
+                return False
+        return self._is_valid_final_value(format_value)
+
+    def _is_valid_final_value(self, format_value: ast.AST) -> bool:
+        # Variable lookup is okay and a single attribute is okay
+        if isinstance(format_value, (ast.Name, ast.Attribute)):
+            return True
+        # Function call with empty arguments is okay
+        elif isinstance(format_value, ast.Call) and not format_value.args:
+            return True
+        # Named lookup, Index lookup & Dict key is okay
+        elif isinstance(format_value, ast.Subscript):
+            if isinstance(format_value.slice, ast.Index):
+                return isinstance(
+                    format_value.slice.value,
+                    self._valid_format_index,
+                )
+        return False
+
+    def _is_valid_chaining(self, format_value: AnyChainable) -> bool:
+        chained_parts: List[ast.AST] = list(attributes.parts(format_value))
+        if len(chained_parts) <= self._max_chained_items:
+            return self._is_valid_chain_structure(chained_parts)
+        return False
+
+    def _is_valid_chain_structure(self, chained_parts: List[ast.AST]) -> bool:
+        """Helper method for ``_is_valid_chaining``."""
+        has_invalid_parts = any(
+            not self._is_valid_final_value(part)
+            for part in chained_parts
+        )
+        if has_invalid_parts:
+            return False
+        if len(chained_parts) == self._max_chained_items:
+            # If there are 3 elements, exactly one must be subscript or
+            # call. This is because we don't allow name.attr.attr
+            return sum(
+                isinstance(part, self._single_use_types)
+                for part in chained_parts
+            ) == 1
+        # All chaining with fewer elements is fine!
+        return True
 
 
 @final